[CELEBORN-718] Celeborn generate hadoop configuration should respect Celeborn conf
